Check SetNodeAffinityArchRequirement() to make sure the nodeSelector does not have an architecture label set.
With the new rules in shouldIgnorePod(). There can be the case where the Required node affinity should not be set, but the check passes because there is preferred node affinity that needs to be set. Thus, a check must be added to prevent adding required node affinity when then NodeSelector has arch lable set